Add order placement feature and logout redirect
Enhanced the application with the ability to place orders. New methods 'placeOrder' and 'checkJwt' were added to productsController and userController, respectively. The 'placeOrder' method checks if an authorized user has enough stock for their requested product quantity and adjusts the stock accordingly. The 'checkJwt' method verifies user's jwt token validity before any operation is carried out. This ensures only authorized and authenticated users can access and modify products.

Also added a redirect in 'index.js' to navigate the user back to the login page if no active user is found in the session storage. This serves as a basic protection against unauthorized access to the application when the user is not logged in.

Changes are also merged into 'index.css' to style new components and HTML elements used for order placement, and error/success messages displays.

File 'server/routes/index.js' is renamed to 'server/routes/products.js' and routes in 'server/app.js' are updated accordingly, for a more semantically meaningful structure.